Refactor to work with recursive subthreads
