+++ src/renderer/components/CommentSection/CommentSection.css
@@ -37,172 +37,11 @@
   }
 }
 
-
-.comment {
-  padding: 15px;
-}
-
 .hideComments {
   font-size: 13px;
   text-decoration: underline;
   cursor: pointer;
   color: var(--title-color);
-}
-
-.commentThumbnail {
-  float: inline-start;
-  inline-size: 60px;
-  block-size: 60px;
-  border-radius: 200px;
-}
-
-.commentThumbnailHidden {
-  float: inline-start;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  inline-size: 60px;
-  block-size: 60px;
-  font-size: 20px;
-  line-height: 1em;
-  color: rgb(0 0 0);
-  background-color: rgb(235 160 172);
-  border-radius: 50%;
-}
-
-.commentAuthorWrapper {
-  font-weight: bold;
-  font-size: 14px;
-  margin-inline-start: 68px;
-  margin-block-start: 0;
-  overflow-wrap: break-word;
-}
-
-.commentOwner {
-  background-color: var(--secondary-card-bg-color);
-  border-radius: 10px;
-  padding-inline: 10px;
-}
-
-
-.commentAuthor {
-  color: inherit;
-  text-decoration: none;
-}
-
-.commentText {
-  white-space: pre-wrap;
-  font-size: 14px;
-  /* stylelint-disable-next-line liberty/use-logical-spec */
-  margin-left: -10px;
-  margin-inline-start: 70px;
-  overflow-wrap: break-word;
-}
-
-/* stylelint-disable liberty/use-logical-spec */
-:global(body[dir='rtl'] .commentText) {
-  margin-left: initial;
-  margin-right: -10px;
-}
-/* stylelint-enable liberty/use-logical-spec */
-
-.commentPinned {
-  font-weight: normal;
-  font-size: 12px;
-  margin-block: 0 5px;
-  margin-inline-start: 68px;
-}
-
-.commentDate {
-  font-weight: normal;
-  margin-inline-start: 5px;
-  font-size: 12px;
-}
-
-.commentMemberIcon {
-  margin-inline-start: 5px;
-  inline-size: 14px;
-  block-size: 14px;
-}
-
-.commentSubscribedIcon {
-  content: var(--logo-icon);
-  block-size: 14px;
-  inline-size: 14px;
-  vertical-align: middle;
-  margin-inline: 0.25em;
-}
-
-.commentLikeCount {
-  font-size: 11px;
-  margin-inline-start: 70px;
-  margin-block-start: 0;
-}
-
-.commentHeartBadge {
-  display: inline-block;
-  position: relative;
-  inline-size: 25px;
-  block-size: 20px;
-  margin-inline-start: 10px;
-  margin-block-end: -7px;
-}
-
-.commentHeartBadgeImg {
-  position: absolute;
-  inset-inline-start: 0;
-  inline-size: 15px;
-  block-size: 15px;
-  border-radius: 50%;
-}
-
-.commentHeartBadgeWhite {
-  position: absolute;
-  inset-inline-start: 9px;
-  inset-block-end: 1px;
-  inline-size: 11px;
-  block-size: 11px;
-  z-index: 1;
-}
-
-.commentHeartBadgeRed {
-  position: absolute;
-  color: var(--red-500);
-  inset-inline-start: 10px;
-  inset-block-end: 2px;
-  inline-size: 9px;
-  block-size: 9px;
-  z-index: 2;
-}
-
-.commentMoreReplies {
-  font-size: 11px;
-  margin-inline-start: 5px;
-  text-decoration: underline;
-  cursor: pointer;
-  color: var(--title-color);
-}
-
-.commentLoadingMoreReplies {
-  font-size: 11px;
-  margin-inline-start: 5px;
-  color: var(--title-color);
-}
-
-.commentReplies {
-  margin-inline-start: 30px;
-}
-
-.showMoreReplies {
-  margin-inline-start: 30px;
-  font-size: 15px;
-  cursor: pointer;
-  text-decoration: underline;
-}
-
-.loadingMoreReplies {
-  margin-inline-start: 30px;
-  font-size: 15px;
 }
 
 .getMoreComments {
