+++ src/renderer/helpers/api/invidious.js
@@ -429,26 +429,44 @@ export async function invidiousGetVideoInformation(videoId) {
  * The complete Triforce, or one or more components of the Triforce.
  * @typedef {object} InvidiousComment
  * @property {string} id
- * @property {string} authorLink
- * @property {string} authorThumb
+ * @property {'invidious'} dataType
  * @property {string} author
+ * @property {string} authorId
+ * @property {string} authorThumb
  * @property {number} likes
  * @property {string} text
- * @property {string} dataType
- * @property {boolean} isOwner
- * @property {boolean} isPinned
- * @property {number} numReplies
- * @property {boolean} hasReplyToken
- * @property {string} replyToken
- * @property {boolean} showReplies
- * @property {InvidiousComment[]} replies
+ * @property {string} time
  * @property {boolean} isHearted
  * @property {boolean} isMember
+ * @property {boolean} isOwner
+ * @property {boolean} isPinned
+ * @property {boolean} hasReplyToken
+ * @property {string} replyToken
  * @property {string} memberIconUrl
- * @property {string} time
+ * @property {number} numReplies
  */
-/** @typedef {{commentCount: number, videoId: string, continuation: string?, comments: InvidiousComment[]}} InvidiousCommentResponse */
-
+/**
+ * See: https://docs.invidious.io/api/#get-apiv1commentsid
+ * @typedef {{
+ *   commentCount: number,
+ *   videoId: string,
+ *   continuation: string?,
+ *   comments: {
+ *     authorId: string,
+ *     author: string,
+ *     authorThumbnail: string,
+ *     authorIsChannelOwner: boolean,
+ *     isSponsor: boolean,
+ *     likeCount: number,
+ *     creatorHeart?: object,
+ *     isPinned: boolean,
+ *     commentId: string,
+ *     contentHtml: string,
+ *     published: number,
+ *     replies?: {replyCount: number, continuation: string}
+ *   }[]
+ * }} InvidiousCommentResponse
+ */
 export async function invidiousGetComments({ id, nextPageToken = '', sortNewest = true }) {
   const payload = {
     resource: 'comments',
@@ -610,24 +628,23 @@ export function invidiousImageUrlToInvidious(url, currentInstance = null) {
 
 /**
  * @param {InvidiousCommentResponse} response
+ * @returns {InvidiousComment[]}
  */
 function parseInvidiousCommentData(response) {
   return response.comments.map((comment) => {
     return {
       id: comment.commentId,
-      authorLink: comment.authorId,
-      authorThumb: youtubeImageUrlToInvidious(comment.authorThumbnails.at(-1).url),
+      dataType: 'invidious',
+      authorId: comment.authorId,
+      authorThumb: youtubeImageUrlToInvidious(comment.authorThumbnail),
       author: comment.author,
       likes: comment.likeCount,
       text: autolinker.link(invidiousImageUrlToInvidious(comment.contentHtml, getCurrentInstanceUrl())),
-      dataType: 'invidious',
       isOwner: comment.authorIsChannelOwner,
       isPinned: comment.isPinned,
       numReplies: comment.replies?.replyCount ?? 0,
       hasReplyToken: !!comment.replies?.continuation,
       replyToken: comment.replies?.continuation ?? '',
-      showReplies: false,
-      replies: [],
       isHearted: comment.creatorHeart !== undefined,
       isMember: comment.isSponsor,
       memberIconUrl: youtubeImageUrlToInvidious(comment.sponsorIconUrl),
